+++ packages/util-crypto/src/address/check.ts
@@ -6,8 +6,8 @@ import { Prefix } from './types';
 
 import bs58 from 'bs58';
 
+import checkChecksum from './checkChecksum';
 import defaults from './defaults';
-import sshash from './sshash';
 
 export default function check (address: string, prefix: Prefix): [boolean, string | null] {
   const decoded = bs58.decode(address);
@@ -18,18 +18,7 @@ export default function check (address: string, prefix: Prefix): [boolean, strin
     return [false, 'Invalid decoded address length'];
   }
 
-  const isPublicKey = decoded.length === 35;
+  const [isValid] = checkChecksum(decoded);
 
-  // non-publicKeys has 1 byte checksums, else default to 2
-  const endPos = decoded.length - (isPublicKey ? 2 : 1);
-
-  // calculate the hash and do the checksum byte checks
-  const hash = sshash(decoded.subarray(0, endPos));
-  const checks = isPublicKey
-    ? decoded[decoded.length - 2] === hash[0] && decoded[decoded.length - 1] === hash[1]
-    : decoded[decoded.length - 1] === hash[0];
-
-  return !checks
-    ? [false, 'Invalid decoded address checksum']
-    : [true, null];
+  return [isValid, isValid ? null : 'Invalid decoded address checksum'];
 }