+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.6.0-beta.x
 
 - Add `signatureVerify (message: Uint8Array | string, signature: Uint8Array | string, addressOrPublicKey: Uint8Array | string): VerifyResult` to verify any signature, regardless of type. Returns validity & detected crypto type.
+- Add `addressCheckChecksum (decoded: Uint8Array)` as an helper to extract and verify the ss58 checksum
 
 # 2.5.1 Feb 22, 2020